--- app/peas.md
@@ -1,146 +1,0 @@
-# A Story as You Like It
-
-## {root}
-
-Do you wish to hear the story of the three alert peas?
-
-[Yes](pod)
-[No](beanpoles)
-
-## {beanpoles}
-
-Would you prefer the story of the three big skinny beanpoles?
-
-[Yes](ablutions)
-[No](bushes)
-
-## {bushes}
-
-Would you prefer the story of the three middling mediocre bushes?
-
-[Yes](bushes_watching)
-[No](finished)
-
-## {pod}
-Once upon a time there were three peas dressed in green who were fast asleep in their pod. Their round faced breathed through the holes in their nostrils, and one could hear their soft and harmonious snoring.
-
-[I prefer another description.](highway)
-[This description suits me.](not_dreaming)
-    
-## {not_dreaming}
-
-They were not dreaming. In fact, these little creatures never dream.
-
-[I prefer that they dream.](dreaming_in_fact)
-[I do not prefer that they dream.](feet)
-
-## {dreaming_in_fact}
-
-They were dreaming. In fact, these little creatures always dream and their nights secrete charming dreams.
-[I wish to know these dreams.](soup)
-[I don't care about them.](feet)
-
-## {feet}
-
-Their cute little feet were covered in warm stockings and in bed they wore black velvet gloves.
-
-[I prefer gloves of another color.](gloves)
-[This color suits me.](proud_mirrors)
-
-## {gloves}
-
-In bed, they wore blue velvet gloves.
-
-[I prefer gloves of another color.](feet)
-[This color suits me.](proud_mirrors)
-
-## {highway}
-
-Once upon a time there were three peas rolling along on the great highway. When evening came, they fell fast asleep, tired and worn.
-
-[I wish to know the rest](not_dreaming)
-[I don't wish to know](finished)
-
-## {proud_mirrors}
-
-All three were dreaming the same dream; indeed, they loved each other tenderly and, like proud mirrors, always dreamed similarly.
-
-[I wish to know their dream.](soup)
-[I don't wish to know it.](opopoi)
-
-## {soup}
-
-They dreamed that they were getting soup at their soup kitchen, and that upon uncovering their bowl they discovered that it was ers soup. Horrified, they woke up.
-
-[I wish to know why they woke up horrified.](websters)
-[I judge it a waste of time to investigate this question further.](opopoi)
-    
-## {websters}
-
-[Consult the word "ers" in Webster's dictionary, and let us hear no more of it.](opopoi)
-
-## {opopoi}
-
-Opopoï! they cried when they opened their eyes. Opopoï! what a dream we dreamed! A bad omen, said the first. Yessir, said the second, that's a fact, and now I'm sad. Don't worry like that, said the third, who was the sharpest of the three. We must comprehend rather than despair, in short, I will analyze it for you.
-[I wish to know the interpretation of this dream right away.](dream_interpretation)
-[I wish on the contrary to know the reactions of the other two.](bored)
-
-## {bored}
-
-You bore us to tears, said the first. Since when do you know how to analyze dreams? Yes, since when? added the second.
-
-[I wish to know since when.](analysis)
-[I don't care.](analysis)
-
-## {analysis}
-
-Since when? cried the third. How should I know? The fact is that I practice analysis. You'll see.
-
-[I wish to see.](dream_interpretation)
-[I do not wish to see.](dream_interpretation)
-
-## {dream_interpretation}
-
-Well then, let's see! cried his brothers. Your irony doesn't please me a bit, replied the other, and you'll not learn a thing. Moreover, during this rather sharp conversation, hasn't your sense of horror been blurred, or even erased? What use then to stir up the mire of your papilionaceous unconscious? Let's rather go wash ourselves in the fountain and greet this gay morning in hygiene and saintly euphoria! No sooner said than done: they slip out of their pod, let themselves roll gently to the ground, and trot joyously to the theatre of their ablutions.
-[I wish to know what happens at the theater of their ablutions.](ablutions)
-[I do not wish to know.](finished)
-
-## {ablutions}
-
-The three big beanpoles were watching them.
-
-[The three big beanpoles displease me.](finished)
-[They suit me.](voyeurized)
-
-## {bushes_watching}
-
-The three middling mediocre bushes were watching them.
-
-[The three middling mediocre bushes displease me.](finished)
-[They suit me.](voyeurized)
-
-## {voyeurized}
-
-Seeing themselves voyeurized in this fashion, the three alert peas, who were very modest, fled.
-
-[I wish to know what they did after that.](back_to_pod)
-[I do not wish to know.](finished)
-
-## {back_to_pod}
-
-They ran very hard back to their pod and, closing the latter after them, went back to sleep.
-
-[I wish to know the rest.](no_rest)
-[I do not wish to know.](finished)
-    
-## {no_rest}
-
-There is no rest and the story is finished. 
-
-THE END
-
-## {finished}
-
-In this case, the story is finished. 
-
-THE END
